Added stats panel to interface
Updated the existing stat_panel and summary table functions to work
with cross filter dimension to report total number of fish, yearling
equivalents, stocking events, and species.  A summary table presenting
the number of events and yearling equivalents stocked by species is
also included. The contents of the summary table and species table
change dynamically in response to the selected filters.