Moved polygon overlay into module
The code associated with the polygon overlays has been moved into a
single module that is instantiated once the data has been loaded.
Three new dimensions were added to main.js reflect the status of the selected
polygons and a function was created to update those dimenstion when
the selected polygons changed.  Most of the remaining changes in this
commit are automatic code changes completed by eslint and prettier.