Added refresh method to Spatial Radio Buttons
Added a refresh method to the spatial radio buttons widget so that it
can be updated through code when the state changes from something
other than a click on the radio button (i.e. - when the map changes
scope). Also fixed a couple of typos in the class neame of the
selected pie in the piechart overlay module that were breaking the
'unhighlight' functionality.